feat(mcp): migrate draft/suggest tools to agent-tools and add activity commit tools
Move record_activity, record_activities, propose_transaction_categories, create_categorization_rule, and prepare_asset_classification from crates/ai (rig tools) into agent-tools as AgentTool impls (Draft/Suggest scopes), preserving names, schemas, and output JSON (parity snapshots). Add MCP-only commit_activity_draft/commit_activity_drafts (Write, requires activities:draft+write) that map ActivityDraft to NewActivity, call the activity service, and clear the health cache. Split the catalog into assistant_catalog() (read + draft/suggest, no commit) and mcp_catalog() (read + draft/suggest + commit); the MCP host defaults to mcp_catalog. The in-app assistant reaches the migrated tools via the rig adapter; commit tools are never exposed to it.
