+++ crates/agent-tools/Cargo.toml
@@ -17,6 +17,7 @@ thiserror = { workspace = true }
 chrono = { workspace = true }
 log = { workspace = true }
 rust_decimal = { workspace = true }
+uuid = { workspace = true }
 
 wealthfolio-core = { path = "../core" }
 wealthfolio-spending = { path = "../spending" }
