+++ crates/ai/src/chat/streaming.rs
@@ -311,23 +311,12 @@ pub(super) async fn spawn_chat_stream<E: AiEnvironment + 'static>(
                 }
             }
 
-            if is_allowed("record_activity") {
-                allowed_tools.push(Box::new(tool_set.record_activity));
-            }
-            if is_allowed("record_activities") {
-                allowed_tools.push(Box::new(tool_set.record_activities));
-            }
+            // record_activity, record_activities, propose_transaction_categories,
+            // create_categorization_rule, and prepare_asset_classification now
+            // come from the shared agent-tools catalog above (via RigAgentTool).
+            // Only import_csv still implements rig's Tool trait directly.
             if is_allowed("import_csv") {
                 allowed_tools.push(Box::new(tool_set.import_csv));
-            }
-            if is_allowed("propose_transaction_categories") {
-                allowed_tools.push(Box::new(tool_set.propose_categories));
-            }
-            if is_allowed("create_categorization_rule") {
-                allowed_tools.push(Box::new(tool_set.create_categorization_rule));
-            }
-            if is_allowed("prepare_asset_classification") {
-                allowed_tools.push(Box::new(tool_set.prepare_asset_classification));
             }
 
             let mut builder = $client
